get rid of Usefield parameters, fix iterationField bug
warning: this is a potentially breaking change if you have an op with the field input connected but that setting switched off.
Note that this also changes the behavior of the combineStairs fields to multiply the radius parameter instead of replacing it and add to the offset parameter instead of replacing it.

see #40