change how validation errors are aggregated
First, go back to including a toolkitVersion column in the definition tables.
That makes it possible for the shaderBuilder to check for mismatches without having to query the parameters of opDefinitions.
Next, rather than having the validation errors defined throughout the network and then pulled together as one, each one aggregates both its own errors and those of its inputs only.
This gets rid of the numerous dependency lines that were showing up in networks.

closes #388