Fix handling of types in multiInputHandler and other related components
see #5.

multiInputHandler now reduces input definitions down to only the types supported by all inputs, and if there are no types supported by all, it produces a validation error.
Consolidated a lot of the type-handling code into a new raytkTypes library so it's easier to keep track of it all.