BREAKING: fixed issue with waveField period scaling
Because the wave "functions" are implemented as macros, when passing an expression into them, the TAU scaling was only being applied to part of the expression.

As a result, waveField now handles periods differently. It looks like it is treating them as 0.5 x how they used to be handled.

close #604