attempted fix for mysterious create/update issue
might be related to the renaming of the iop*shortcut etc params.
for pennerEasingFn, for some reason when it's created by the palette or updated, it puts the expression from the `Name` custom par on the opDefinition into `iop1shortcut`.
seems to only happen for that operator.
this change is removing unused iop params from opDefinition